implement suggestions and fix tests
