+++ jenkins-design-language/LICENSE
@@ -1,6 +1,6 @@
 The MIT License
 
-Copyright (c) 2016 CloudBees Inc and a number of other of contributors
+Copyright (c) 2017 CloudBees Inc and a number of other of contributors
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
