+++ jenkins-design-language/package.json
@@ -60,7 +60,7 @@
     "react-addons-css-transition-group": "15.4.2",
     "react-dom": "15.4.2",
     "react-router": "2.8.1",
-    "@jenkins-cd/blueocean-material-icons": "0.0.5"
+    "@jenkins-cd/blueocean-material-icons": "0.0.6"
 
   },
   "devDependencies": {
